% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_stats_plots.R
\name{generate_stats_plots}
\alias{generate_stats_plots}
\alias{generate_stats_plot}
\title{Generate}
\usage{
generate_stats_plot(
  file_path,
  x_var,
  y_var,
  spec,
  x_label,
  y_label,
  plot_title
)
}
\arguments{
\item{t}{[integer] time passed. Can be a single integer or a vector of integers.}

\item{lambdas}{[vector] a vector of failure rates for components, named $\lambda$}

\item{type}{[character] a single value describing whether these probabilities should be combined using the rules of series or parallel systems.}
}
\description{
The function takes in input about the solar power data with the parameters of interest being
solar power output, the time frame for which the data was collected, temperature, pressure,
wind speed, wind direction. With this data we can either view the statistical process control
}
\note{
- `@name` 
- `@title` is the official title and can be anything.
- `@description` is a description that can span multiple lines
- `@param` marks each input parameters for your function.
- `@export` exports your function to be available to a user of your package. (Eg. not an internal function)
- `@importFrom` bundles into your package 1 or more specific functions from another package, so that your package will always function.

You can specify default inputs for an input parameter like with `type = "series"` below.
}
\examples{

# Get series system probability at each time t
get_prob(t = c(2,4,5,6), lambdas = c(0.001, 0.02), type = "series")

# Get parallel system probability at each time t
get_prob(t = c(2,4,5,6), lambdas = c(0.001, 0.02), type = "parallel")

}
\author{
Geet Chheda, Sagar Shenoy, Animesh Chaturvedi, Ishwari Joshi, Shikhar Singh
}
