% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_stats_plots.R
\name{generate_stats_plots}
\alias{generate_stats_plots}
\alias{generate_stats_plot}
\title{Generate}
\usage{
generate_stats_plot(
  file_path,
  x_var,
  y_var,
  spec,
  x_label,
  y_label,
  plot_title
)
}
\arguments{
\item{file_path}{[.csv] Add a .csv file for your function to read}

\item{x_var}{[character] Input column name of the independent variable - 
If using statistical process control with time (date_time) has to be the input variable 
with a start format of YYYY-MM-DD H:M:SS 
If understanding the impact then just use the column name}

\item{y_var}{[character] Output column which will always be power in most cases}

\item{spec}{[numeric] The specification limit i.e. the minimum power required for the local area}

\item{x_label}{[character] label of x-axis for the plot to be generated}

\item{y_label}{[character] label of y-axis for the plot to be generated}

\item{plot_title}{[character] title for the plot to be generated}
}
\description{
The function takes in input about the solar power data with the parameters of interest being
solar power output, the time frame for which the data was collected, temperature, pressure,
wind speed, wind direction. With this data we can either view the statistical process control
}
\note{
You can specify default inputs for an input parameter like with `type = "series"` below.
}
\author{
Geet Chheda, Sagar Shenoy, Animesh Chaturvedi, Ishwari Joshi, Shikhar Singh
}
